refactor token
