+++ crabml-core/src/tokenizer/mod.rs
@@ -1,3 +1,4 @@
-mod bpe;
+mod tokenizer;
+mod tokenizer_llama;
 
-pub use bpe::BpeTokenizer;
+pub use tokenizer::BpeTokenizer;
